requirements.txt
amqp==5.3.1
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
async-timeout==5.0.1
autopep8==2.0.1
billiard==4.2.1
Brotli==1.1.0
celery==5.4.0
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
crispy-bootstrap5==0.7
cron-descriptor==1.4.5
cryptography==42.0.0
cssselect2==0.7.0
distlib==0.3.6
dj-database-url==2.1.0
Django==5.1.5
django-allauth==0.62.1
django-appconf==1.0.6
django-celery-beat==2.7.0
django-composite-foreignkey==1.1.0
django-crispy-forms==2.0
django-filter==24.2
django-jazzmin==3.0.0
django-model-utils==4.5.0
django-notifications-hq==1.8.3
django-oauth-toolkit==2.3.0
django-select2==8.1.2
django-timezone-field==7.1
et_xmlfile==2.0.0
filelock==3.9.0
fonttools==4.47.2
freetype-py==2.3.0
gunicorn==22.0.0
html5lib==1.1
idna==3.6
jsonfield==3.1.0
jwcrypto==1.5.6
kombu==5.4.2
lxml==5.1.0
mod-wsgi==5.0.2
oauthlib==3.2.2
openpyxl==3.1.5
oscrypto==1.3.0
packaging==24.0
pillow==10.2.0
platformdirs==3.0.0
prompt_toolkit==3.0.50
psycopg2==2.9.5
psycopg2-binary==2.9.9
pycairo==1.25.1
pycodestyle==2.10.0
pycparser==2.21
pydyf==0.11.0
pyHanko==0.21.0
pyhanko-certvalidator==0.26.3
pypdf==4.0.0
PyPDF3==1.0.6
pyphen==0.14.0
pypng==0.20220715.0
python-bidi==0.6.6
python-crontab==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
qrcode==7.4.2
redis==5.2.1
reportlab==4.0.9
requests==2.31.0
rlPyCairo==0.3.0
six==1.16.0
sqlparse==0.4.3
svglib==1.5.1
swapper==1.3.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2022.7
tzlocal==5.2
uritools==4.0.2
urllib3==2.1.0
vine==5.1.0
virtualenv==20.20.0
waitress==3.0.1
wcwidth==0.2.13
weasyprint==64.0
webencodings==0.5.1
whitenoise==6.6.0
zopfli==0.2.3
